:recycle: (weather) tidy up weather command
