+++ Commands/Weather/Domain/WeatherBeaconEntity.cs
