+++ Commands/Weather/Domain/WeatherEntity.cs
