+++ Commands/Weather/Domain/WeatherType.cs
