+++ Commands/Weather/Presenter/WeatherFormatter.cs
